Move DrawableShaded trait to kas_theme; add DrawableShadedExt
