+++ kas-theme/Cargo.toml
@@ -15,6 +15,12 @@ documentation = "https://docs.rs/kas-theme/"
 default = ["stack_dst"]
 nightly = ["unsize", "kas/nightly"]
 
+# Enables documentation of APIs for shells and internal usage.
+# This API is not intended for use by end-user applications and
+# thus is omitted from built documentation by default.
+# This flag does not change the API, only built documentation.
+internal_doc = ["kas/internal_doc"]
+
 # Enable config read/write
 config = ["serde", "kas/config"]
