+++ kas-theme/src/lib.rs
@@ -21,6 +21,7 @@
 mod colors;
 mod config;
 mod dim;
+mod draw_shaded;
 mod flat_theme;
 #[cfg(feature = "stack_dst")]
 mod multi;
@@ -34,6 +35,7 @@ pub use kas;
 pub use colors::{Colors, ColorsLinear, ColorsSrgb};
 pub use config::{Config, RasterConfig};
 pub use dim::{Dimensions, DimensionsParams, DimensionsWindow};
+pub use draw_shaded::{DrawableShaded, DrawableShadedExt};
 pub use flat_theme::FlatTheme;
 #[cfg(feature = "stack_dst")]
 pub use multi::{MultiTheme, MultiThemeBuilder};
